# src/main/java/io/github/lyes_sefiane/grocery_items_management/exception/entities/ErrorResponse.java
package io.github.lyes_sefiane.grocery_items_management.exception.entities;

/**
 * Project Name : grocery-items-management-application
 * Package : com.lyess.groceryitems.exception.entities
 * Class : ErrorResponse.java
 * User : Lyes Sefiane
 * Created : 2023-09-29 11:29 AM
 */
public interface ErrorResponse {
    //
}